+++ test/oi/topology/graph_test.exs
@@ -12,7 +12,7 @@ defmodule Oi.Topology.GraphTest do
       node_1 = %Node{id: :node_1, container: DummyStep1, inputs: [:in], outputs: [:out]}
       node_2 = %Node{id: :node_2, container: DummyStep1, inputs: [:in], outputs: [:out]}
 
-      assert Map.from_struct(node_1) != Map.from_struct(node_2)
+      refute Map.equal?(Map.from_struct(node_1), Map.from_struct(node_2))
 
       {:ok, res1} = node_1.container.run(%Orchid.Param{payload: "NodeIn"}, [])
       {:ok, res2} = node_2.container.run(%Orchid.Param{payload: "NodeIn"}, [])